feat: introduce NPM package and basic schema
I know it's a little early to commit a schema but the API of the library to me,
has to be very intentional and ergonomic both for lower-level integrations like
debugging, logging, persistance, etc. and for higher-level programs like
common web applications today.

I think being mindful of the API upfront also gives a sort of _target_ when
answering questions like "how will we store actor messages?" or "how will
applications model request/response async patterns?", etc.

I've chosen to declare a global `Actors` namespace as a convenience for
accessing the types from JS Doc comments. If this poses a serious collision
problem, we can always migrate to using imports the type annotations.