feat: basic local implementation
So here's a first pass at spawn actors _locally_ in the browser. _Locally_ here
means that all spawned actors are running in the same browser context. If the
library is imported in the main thread, all actors will be spawned in the main
thread. If the library is imported in a worker, all actors will be spawned in
that worker. If the library is imported in Node, all actors will be spawned on
the main thread.

Mailboxes (a sequence of incoming messages addressed to a single actor) are
implemented as a simple array. This is because the mailbox is essentially a
queue and actors handling messages sequentially is guaranteed by the event
loop and the recursive implementation used to _enqueue_ an actor for message
processing time.

Scheduling is very primitive at the moment. When a message arrives in the
actor's mailbox, the actor is _enqueued_ in the event loop to eventually
process that message. If this actor still has messages in its mailbox after
processing the current message, it will be _enqueued_ again to process the
next message, and so on until the mailbox is empty.

Caveat, most robust implementations of the actors isolate memory and execution
of actors from each other. To avoid having a few actors monopolize the
available memory or compute time. This implementation does not have that as
JavaScript does not provides primitives in the browser to manage memory or
compute time so granularly. Workers are realistically the most effective way
of separating compute time and memory management isn't easily possible. These
points are important but not critical yet for a basic implementation.