chore: add GPL-3.0 license file
I'm not stuck to this choice. At the moment, I've chosen GPL-3.0 because it's
a license many are afraid to use in their codebases without understanding it
so it seems less likely for anyone to randomly copy this code carelessly and
start using it.

I'll likely relax this in the near future.